ahi vamos con la vista de datos de student

+++ app/src/main/java/com/example/actividad/studentDataBase/model/Student.kt
@@ -17,5 +17,5 @@ data class Student(
     @ColumnInfo(name = "grupo")
     val grupo: Char,
     @ColumnInfo(name = "promedio")
-    val puntaje: Float
+    val puntaje: Double
 )
